[production_env][refactor] refactor main tests dir -> now we have integration with external API, precision and unit
