+++ tests/gower_api/precision/test_handle_unseen_categorical_ordinal.py
