+++ tests/gower_api/unit/test_config_validators.py
@@ -1,18 +1,21 @@
 """Tests for Config field validators — covers every validation error branch."""
 
+import numpy as np
 import pytest
 from pydantic import ValidationError
 
 from gower_metric import Config
 
+DEFAULT_DTYPE = np.float64
+
 
 class TestFeatureTypeValidation:
     def test_invalid_feature_type_raises(self) -> None:
         with pytest.raises(ValidationError, match="Invalid feature type"):
-            Config(feature_types={0: "unknown_type"})
+            Config(feature_types={0: "unknown_type"}, data_type=DEFAULT_DTYPE)
 
     def test_valid_feature_types_pass(self) -> None:
-        cfg = Config(feature_types={0: "numeric"})
+        cfg = Config(feature_types={0: "numeric"}, data_type=DEFAULT_DTYPE)
         assert cfg.feature_types[0] == "numeric"
 
 
@@ -24,6 +27,7 @@ class TestScaleWindowTypeValidation:
                 feature_types={0: "numeric"},
                 scale_window=None,
                 scale_window_type="silverman",
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_valid_scale_window_type_with_kde(self) -> None:
@@ -31,6 +35,7 @@ class TestScaleWindowTypeValidation:
             feature_types={0: "numeric"},
             scale_window="kde",
             scale_window_type="silverman",
+            data_type=DEFAULT_DTYPE,
         )
         assert cfg.scale_window_type == "silverman"
 
@@ -39,6 +44,7 @@ class TestScaleWindowTypeValidation:
             feature_types={0: "numeric"},
             scale_window="kde",
             scale_window_type=None,
+            data_type=DEFAULT_DTYPE,
         )
         assert cfg.scale_window_type is None
 
@@ -49,14 +55,18 @@ class TestKNeighborsValidation:
             ValidationError,
             match="k_neighbors must be None or a positive integer",
         ):
-            Config(feature_types={0: "numeric"}, k_neighbors=0)
+            Config(feature_types={0: "numeric"}, k_neighbors=0, data_type=DEFAULT_DTYPE)
 
     def test_negative_raises(self) -> None:
         with pytest.raises(
             ValidationError,
             match="k_neighbors must be None or a positive integer",
         ):
-            Config(feature_types={0: "numeric"}, k_neighbors=-5)
+            Config(
+                feature_types={0: "numeric"},
+                k_neighbors=-5,
+                data_type=DEFAULT_DTYPE,
+            )
 
 
 class TestOrdinalOrderValidation:
@@ -65,6 +75,7 @@ class TestOrdinalOrderValidation:
             Config(
                 feature_types={0: "categorical_ordinal"},
                 categorical_ordinal_values_order=None,
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_partial_order_missing_column_raises(self) -> None:
@@ -72,6 +83,7 @@ class TestOrdinalOrderValidation:
             Config(
                 feature_types={0: "categorical_ordinal", 1: "categorical_ordinal"},
                 categorical_ordinal_values_order={0: ["a", "b"]},
+                data_type=DEFAULT_DTYPE,
             )
 
 
@@ -86,6 +98,7 @@ class TestHandleUnseenValidation:
             Config(
                 feature_types={0: "binary_asymmetric"},
                 handle_unseen_binary_asymmetric="bad",  # type: ignore[arg-type]
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_invalid_handle_unseen_binary_symmetric(self) -> None:
@@ -93,6 +106,7 @@ class TestHandleUnseenValidation:
             Config(
                 feature_types={0: "binary_symmetric"},
                 handle_unseen_binary_symmetric="bad",  # type: ignore[arg-type]
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_invalid_handle_unseen_categorical_nominal(self) -> None:
@@ -100,6 +114,7 @@ class TestHandleUnseenValidation:
             Config(
                 feature_types={0: "categorical_nominal"},
                 handle_unseen_categorical_nominal="bad",  # type: ignore[arg-type]
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_invalid_handle_unseen_categorical_ordinal(self) -> None:
@@ -108,6 +123,7 @@ class TestHandleUnseenValidation:
                 feature_types={0: "categorical_ordinal"},
                 categorical_ordinal_values_order={0: ["a", "b"]},
                 handle_unseen_categorical_ordinal="bad",  # type: ignore[arg-type]
+                data_type=DEFAULT_DTYPE,
             )
 
 
@@ -117,6 +133,7 @@ class TestBinaryValueOrderValidation:
             Config(
                 feature_types={0: "binary_asymmetric"},
                 binary_asymmetric_value_order={0: [0, 1, 2]},
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_asymmetric_duplicate_values_raises(self) -> None:
@@ -124,6 +141,7 @@ class TestBinaryValueOrderValidation:
             Config(
                 feature_types={0: "binary_asymmetric"},
                 binary_asymmetric_value_order={0: [1, 1]},
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_asymmetric_extra_column_raises(self) -> None:
@@ -131,6 +149,7 @@ class TestBinaryValueOrderValidation:
             Config(
                 feature_types={0: "binary_asymmetric"},
                 binary_asymmetric_value_order={0: [0, 1], 1: [0, 1]},
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_symmetric_wrong_count_raises(self) -> None:
@@ -138,6 +157,7 @@ class TestBinaryValueOrderValidation:
             Config(
                 feature_types={0: "binary_symmetric"},
                 binary_symmetric_value_order={0: [0, 1, 2]},
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_symmetric_duplicate_values_raises(self) -> None:
@@ -145,6 +165,7 @@ class TestBinaryValueOrderValidation:
             Config(
                 feature_types={0: "binary_symmetric"},
                 binary_symmetric_value_order={0: [1, 1]},
+                data_type=DEFAULT_DTYPE,
             )
 
     def test_symmetric_extra_column_raises(self) -> None:
@@ -152,6 +173,7 @@ class TestBinaryValueOrderValidation:
             Config(
                 feature_types={0: "binary_symmetric"},
                 binary_symmetric_value_order={0: [0, 1], 1: [0, 1]},
+                data_type=DEFAULT_DTYPE,
             )
 
 
@@ -162,4 +184,5 @@ class TestConditionalDistancesValidation:
             Config(
                 feature_types={0: "numeric", 1: "categorical_nominal"},
                 conditional_distances=2,  # type: ignore[arg-type]
+                data_type=DEFAULT_DTYPE,
             )
